pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-hub"
version = "0.1.0"
description = "Linux desktop app for monitoring Claude Code and Codex CLI sessions"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "websockets>=13.0",
    "watchdog>=5.0.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "pywebview>=5.3",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pyright>=1.1.390",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
]
tray = [
    "pystray>=0.19.5",
    "Pillow>=11.0.0",
]

[project.scripts]
agent-hub = "agent_hub.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_hub"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
include = ["src"]
exclude = ["tests"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
